api/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1

# MinIO
minio==7.2.0

# Celery
celery==5.3.6

# Authentication
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# ID generation
python-ulid==2.2.0

# Logging
structlog==24.1.0

# Audio processing
pydub==0.25.1

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
